Define a clear separation between Babel and CLI options.
